menambahkan cart

+++ hapus_produk.php
@@ -2,16 +2,23 @@
 session_start();
 include 'admin/koneksi.php';
 
-$response = ['success' => false];
+$response = ['success' => false, 'message' => 'Terjadi kesalahan'];
+
+if (!isset($_SESSION['id_user'])) {
+    $response['message'] = 'Anda harus login untuk menghapus produk dari keranjang';
+    header('Content-Type: application/json');
+    echo json_encode($response);
+    exit;
+}
 
 if (isset($_POST['id'])) {
-    $id_pesanan = $_POST['id'];
-    $id_user = $_SESSION['id_user'];
+    $id_pesanan = mysqli_real_escape_string($koneksi, $_POST['id']);
+    $id_user = mysqli_real_escape_string($koneksi, $_SESSION['id_user']);
 
     // Delete the item from the cart
     $query = mysqli_query($koneksi, "DELETE FROM tb_pesanan WHERE id_pesanan = '$id_pesanan' AND id_user = '$id_user'");
 
-    if ($query) {
+    if ($query && mysqli_affected_rows($koneksi) > 0) {
         // Recalculate totals
         $query_pesanan = mysqli_query($koneksi, "
             SELECT p.*, pr.harga 
@@ -21,8 +28,10 @@ if (isset($_POST['id'])) {
         ");
 
         $subtotal = 0;
+        $item_count = 0;
         while ($row = mysqli_fetch_assoc($query_pesanan)) {
             $subtotal += $row['qty'] * $row['harga'];
+            $item_count++;
         }
 
         $diskon = 0;
@@ -35,11 +44,17 @@ if (isset($_POST['id'])) {
 
         $response = [
             'success' => true,
+            'message' => 'Produk berhasil dihapus dari keranjang',
             'subtotal' => $subtotal,
             'diskon' => $diskon,
-            'total_bayar' => $total_bayar
+            'total_bayar' => $total_bayar,
+            'item_count' => $item_count
         ];
+    } else {
+        $response['message'] = 'Produk tidak ditemukan atau sudah dihapus';
     }
+} else {
+    $response['message'] = 'ID pesanan tidak diberikan';
 }
 
 header('Content-Type: application/json');
